Replace vim's auto-linebreak with highlighting of the 80th column.
This might be more preferable since it does not automatically break code or enforces you to line-break there. Sometimes it might be preferable to have lines longer than 80 characters, for example when quickly sketching out an idea. I guess this is more in the spirit of PEP8, where having a line length of 79 characters is just a recommendation and no enforcement.

Add Syntax Highlighting to the vim/emacs code snippets.