Remove comment annotations
This is the only subsection of the guide,
which is perhaps a smell.

Its recommendations conflict with
earlier recommendations in the "Comments" section:

> Write self-documenting code and ignore the rest of this section.
> Avoid writing comments to explain bad code.
> Refactor the code to make it self-explanatory.
> Keep existing comments up-to-date.
> An outdated comment is worse than no comment at all.

Each of `TODO`, `FIXME`, `OPTIMIZE`, `HACK`, and `REVIEW`
promote bad habits,
which can be better addressed by the earlier "Comments" section
or by using a project management system,
benchmarking tools,
or code review tools.