Convert to AsciiDoc
AsciiDoc allows us to have native admonitions, footnotes, sane nesting
of list items, etc and to easily export the guide in various formats.
That would also make it trivial to publish a simple site from the guide
down the road if we want to.

A longer term plan is to convert all guides to AsciiDoc and maybe even
RuboCop's own documentation. RSpec and Rails style guides have been recently
converted:
  - https://github.com/rubocop-hq/rspec-style-guide/pull/97
  - https://github.com/rubocop-hq/rails-style-guide/pull/250