Add `when` description/example to `no-then` rule
There's a PR to add a cop named `MultilineWhenThen`,
which detects `then` in multiline `when`.
https://github.com/rubocop-hq/rubocop/pull/7114
It refers to `MultilineIfThen` cop, which is based on `no-then` rule here.
So this commit adds `when` description and example to `no-then` rule,
making newly added `MultilineWhenThen` cop based on the style guide.
FYI, the discussion about this feature is here: https://github.com/rubocop-hq/rubocop/issues/7064